pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forecast-baseline"
version = "0.1.0"
description = "Minimal baseline for FORECAST task (train/predict/evaluate)."
authors = [{ name = "Baseline" }]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "nltk>=3.9",
]

[project.scripts]
forecast-baseline = "src.app.cli:main"
